+++ CHANGELOG.md
@@ -15,6 +15,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Upgrade to Holochain v0.0.170
 
+### Fixed
+- Fix AppEntryType type definition, leading to wrong deserialization.
+
 ## 2022-09-30: v0.9.2
 ### Fixed
 - Response type for request `RestoreCloneCell` set to `InstalledCell`.